[S09P22B310-196] feat : 밸런스게임 기능 추가

+++ backend/src/main/java/com/ntt/mwonimoney/domain/game/model/dto/BalanceGameHistoryDto.java
@@ -11,5 +11,7 @@ public class BalanceGameHistoryDto {
 
 	private Long memberIdx;
 
-	private byte selectAnswer;
+	private int selectAnswer;
+
+
 }
